Tests/LibWeb/Text/expected/hit_testing/svg-fill-geometry.txt

inside triangle fill: triangle
triangle bounding box corner outside fill: svg
inside unfilled rect: svg
